Update support for sending new shared files to IRC.
The message format for shared files changed a while back, and broke for
a while. It seems fairly reliable now, as long as we're using the
correct payload. The downside is that, as of today actually, it's not
possible to get new user API tokens needed for the API to make files
public, but that's just how it goes.

This change updates the code for the new message payload for shared
files and to fix stuff some around blank messages appearing during an
upload.